feat: section8 - 15 수강완료

+++ Node/nomad_youtube_clone/wetube/src/controllers/videoController.js
@@ -20,20 +20,31 @@ export const watch = async (req, res) => {
 }
 export const getEdit = async (req, res) => {
   const { id } = req.params
+  const { user: { _id } } = req.session
   const video = await Video.findById(id)
   if (!video) {
     return res.status(404).render('404', { pageTitle: 'Video not found' })
   }
 
+  if (String(video.owner) !== String(_id)) {
+    return res.status(430).redirect('/')
+  }
+
   return res.render('edit', { pageTitle: `Editing: ${video.title}`, video })
 }
 export const postEdit = async (req, res) => {
+  const { user: { _id } } = req.session
   const { id } = req.params
   const { title, description, hashtags } = req.body
   const video = await Video.exists({ _id: id })
   if (!video) {
     return res.status(404).render('404', { pageTitle: 'Video not found' })
   }
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(430).redirect('/')
+  }
+
   await Video.findByIdAndUpdate(id, {
     title,
     description,
@@ -73,7 +84,18 @@ export const postUpload = async (req, res) => {
 }
 
 export const deleteVideo = async (req, res) => {
+  const { user: { _id } } = req.session
   const { id } = req.params
+  const video = await Video.findById(id)
+
+  if (!video) {
+    return res.status(404).render('404', { pageTitle: 'Video not found' })
+  }
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(430).redirect('/')
+  }
+
   await Video.findByIdAndDelete(id)
   return res.redirect('/')
 }
